Rename pkg to internal. Closes #14
