+++ internal/plugins/manage.go
