+++ internal/plugins/plugin_spec_test.go
